fix: Actually center build with, for, on blocks
Before because the width of the blocks was set to 80% of the parent,
and I was setting large left margins, the blocks could never truly be
centered because of the parent would expand to hold the new left margin,
then the child would expand to be 80% of that, and the parent would
expand to hold the child and so on.

I fixed this by setting the child (block) to have a fixed witdh.