refactor: Move page components into pages directory
I couldn't do this before because next.js automatically
creates pages for all .js files at any level within the pages
directory, treating subdirectories as paths in the url.

But I discovered that if you prefix the .js file name with an
underscore, then those paths aren't created!

And this is a much less awkward place for the page components
to live, instead of within the compoenents directory, which really
makes things hard to find in there.