fix: Make cards split horizontally on small screens
The image still doesn't look great often times, but it's better
than the alternatives of leaving it as it was where the text gets
all kinds of funky, or to 'contain' the image, leaving it tiny
and awkward.

You can't always see the entire image on smaller screens, but they're
just meant to be pretty, the text is the key component of a card.