fix: Card size when text is small
There was a problem with the width of the card when the
card was split horizontally, and the text wasn't long
enough to stretch out the card.  The card condenced into
the center.

Because I'm using the Mui button base to get the ripple
onto my card, I also have to deal with their styles.  One
of them is justify-content: center, which is supposed to
affect their display: flex, but was also affecting my
display: grid.

The solution was just to override that with
justify-content: stretch.