chore: Use optimized version of images
Even though the optimization build process is in place, if you
specify the path to the images directly, it will use the raw
image instead of the optimized one.  To use the optimized one,
the image simply has to be require()'d, which will produce a
string path to the optimized image.