refactor: Move files into new src/ and public/ dirs
Next 9.1 now FINALLY supports the src directory!  So I
can finally put my code in there without splitting it up
in a weird way.

They also support a public/ directory, which will be served
from the root.  Because of that, they're deprecating static/,
and instead changing it to public/static/.